+++ batch/src/App.js
@@ -2,8 +2,8 @@ import React, { Component } from 'react';
 import 'bootstrap/dist/css/bootstrap.css';
 import './App.css';
 //import {comp_test_data} from './testdata';
-import Navigation from './components/Navigation';
-import Content from './components/Content';
+import Navigation from './Components/Navigation';
+import Content from './Components/Content';
 import Auth from './Auth/Auth';
 
 export default class App extends Component {
